+++ files_sdk/models/action.py
@@ -16,7 +16,7 @@ class Action:
         "display": None,  # string - Friendly displayed output
         "ip": None,  # string - IP Address that performed this action
         "source": None,  # string - The source path for this action, if applicable
-        "targets": None,  # array - Targets
+        "targets": None,  # array(object) - Targets
         "user_id": None,  # int64 - User ID
         "username": None,  # string - Username
         "action": None,  # string - Type of action