+++ files_sdk/models/notification.py
@@ -14,8 +14,8 @@ class Notification:
         "path": None,  # string - Folder path to notify on This must be slash-delimited, but it must neither start nor end with a slash. Maximum of 5000 characters.
         "group_id": None,  # int64 - ID of Group to receive notifications
         "group_name": None,  # string - Group name, if a Group ID is set
-        "triggering_group_ids": None,  # array - If set, will only notify on actions made by a member of one of the specified groups
-        "triggering_user_ids": None,  # array - If set, will onlynotify on actions made one of the specified users
+        "triggering_group_ids": None,  # array(int64) - If set, will only notify on actions made by a member of one of the specified groups
+        "triggering_user_ids": None,  # array(int64) - If set, will onlynotify on actions made one of the specified users
         "trigger_by_share_recipients": None,  # boolean - Notify when actions are performed by a share recipient?
         "notify_user_actions": None,  # boolean - If true, will send notifications about a user's own activity to that user.  If false, only activity performed by other users (or anonymous users) will be sent in notifications.
         "notify_on_copy": None,  # boolean - Trigger on files copied to this path?
@@ -26,7 +26,7 @@ class Notification:
         "recursive": None,  # boolean - Apply notification recursively?  This will enable notifications for each subfolder.
         "send_interval": None,  # string - The time interval that notifications are aggregated to
         "message": None,  # string - Custom message to include in notification emails
-        "triggering_filenames": None,  # array - Array of filenames (possibly with wildcards) to scope trigger
+        "triggering_filenames": None,  # array(string) - Array of filenames (possibly with wildcards) to scope trigger
         "unsubscribed": None,  # boolean - Is the user unsubscribed from this notification?
         "unsubscribed_reason": None,  # string - The reason that the user unsubscribed
         "user_id": None,  # int64 - Notification user ID